+++ modules/shared/core/src/sap/commerce/toolset/Plugin.kt
@@ -46,7 +46,8 @@ enum class Plugin(val id: String, val url: String? = null) {
     JAVAEE_EL("com.intellij.javaee.el", "https://plugins.jetbrains.com/plugin/20208-jakarta-ee-expression-language-el-"),
     SPRING("com.intellij.spring", "https://plugins.jetbrains.com/plugin/20221-spring"),
     CRON("com.intellij.cron", "https://plugins.jetbrains.com/plugin/24438-cron-expressions"),
-    GRID("intellij.grid.plugin");
+    GRID("intellij.grid.plugin"),
+    JAVA_I18N("com.intellij.java-i18n");
 
     val pluginId: PluginId
         get() = PluginId.getId(id)